pyproject.toml
[project]
name = "note-writer-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pytest>=9.0.1",
]

[project.scripts]
note-writer-lab = "note_writer_lab.cli:main"
